Working on savegame serialization
It gets serialized up to single gameobjects now.
Still left is the list of GODataBlocks to represent a full scene.

Lastly there should be a header file describing the current scene of the game
and a separate player position.
This will then contain a reference to the DataBlock list of the curretn scene.
These information should be used to restore a savegame in the end.