Implemented full data model for save game
The last layer of the model is now a SaveFile which is able
to store any IDataBlock by name.
IDataBlock must be ProtoContract decorated or else things will fail.

Also created a user for the whole saving thing and implemented events
that can be thrown in order to independently issue saves and loads.