Finishing CSP Editor & Cleaning
Renamed some CSP related members so it's clearer what exactly they are good for.
Removed the edit button from the csp editor window.
It's really not required. Moving input information from one output
to the other might not be a trivial thing, I shall investigate,
if it can be done on the fly or if it will stay the way it is.
(You cannot change existing output methods and retain the input information)