Changing how data context in models is generated
ModelTasks must be annotated with the ModelDataDescriptorAttribute.
This provides information about type, display name and context relevant
fields.

Context relevant fields are retrived from the ExecutorTask that
corresponds to the model.
It will take fields annotated with the FieldDefinitionAttribute.
This attribute contains information about how to render and manage the
field.