Worked out the rendering process a little
Started on displaying the sequencer window and execution type etc.

Implemented behaviours and stuff to do that.

Added extra sorting delta to children to sort out issues
with children in default renderer getting displayed wrong.

Fixed a few things on the way.

This stuff ain't working properly yet.
It errors hard when trying to remove a child object in
ModelStatusResponder. I assume it has to do with values not getting set
properly, i'll see about it later.