Fixing indeting problems
Indents are now displayed correctly, albeit the buttons look awful atm.
Should create a gui skin for the editor.

I ran a couple of tests with serialization / deserialization,
everything seems to be in order but before this can be deemed production
ready, I think the behaviour should be monitored a little more.

Just in case a quirk creeps in.