Working on build process optimisation
Okay so the guiskin is a unitypackage now,
which is required in order to keep the image links intact.

Also I changed the VS post-build a little to account for the new
make_distribution script which will take some arguments and generate a
unitypackage automatically.

Needs some testing I suppose.