Performance and usability considerations for SequenceExecutor
Cut down the ExecutionType selection to 2 things.
AUTO and TRIGGER.

_ONCE options got removed because:
Generally once should mean that you want to evolve through the tree
exactly one time. Which is the default behaviour. If you want stuff to
continue, you wrap it all up in a repeater operator.

Removed the _PARALLEL connotation because the whole thing runs parallel
anyway. It never actually stopped the whole game progress, which is the
meaning of "not parallel".

Additionally the tree is now executed in a coroutine which should ease up
the CPU load a little bit by removing the overhead of calling Update()
and also by parallelising (that the word) the execution a little.

If this comes with some severe implications I might make a sequencer
executor version that runs in the Update() method again for these cases.