Saving BT structure in SequenceExecutor now works
Removed the SequenceContainer and moved this all to SequenceExecutor.
An intermediate type was really not required.

Forcing a scene save now everytime the operator editor window is closed.
This is a dirty friggin hack to force unity to persist changes in the
serialized BT data properly. It's probably a bug with the unity
serializer. Which also means it's likely not going to be fixed any time
soon.