Removed ThreadManager & Entity & Mass-Formatting
ThreadManager was only used in OutputEventListener and just because
there was, at some point in time and space, no reference to a
MonoBehavior.
Since there really is one now (the SignalProcessor which owns the invoked
saignal) I'm using this. Makes a lot of sense and is nicer & cleaner.

Removed the Logic namespace and its content (Entity) - not sure what that was about.