Allow serializing / saving of non-scalar values
Trimmed some fat from FieldDescriptor.

During creating / restoring saves, serialized data from primitives is now
always interpreted as object. That allows for serializing scalar values as
well as array values. Also this avoids some unboxing of primitives when
deserializing and passing that on to fasterflect which boxes it up again.
It now stays in its box.