Menu for quickly adding CSP objects to scenes
Created a menu Entry (Playblack/Csp).
This will read a FGD file from Assets/Playblack/csp_definitions.fgd
and generate context menu entries for adding CSP objects that are
defined in that FGD file.

The csp_definitions.fgd is a json file containing an array/list
with these elements:

Category: Can be a word or a / separated thing. It's used for grouping the
menu entries.

NameInMenu: The display name in context menu and of the geenrated GO

TypeName: FQ Class name of the component that makes up the CSP

Components: An array with components to resolve
dependencies. Can also be used to just add multiple CSP objects and
also resolve their dependencies.