Changes and bulletproofing in savegame processing
Added a try/catch block around the serialization process
in SaveManager so that it doesn't explode if only on field
fails to be serialized.

Also added a property "persists on restore".
That is a flag a save state can check and not destroy
the savemanagers GO on restore but inject the saved data
into the existing object instead.

Should be used only on objects that are not destroyed during runtime.