SaveManager added storeFullObjectTree flag
When true this will store the complete object tree (which is the old behaviour)
This is useful when you're working with prefabs.

When false the SaveManager will only scan for saveable components on the
current game object.
This is useful when you're working with loose object structures
where there are multiple save managers in the hierarchy
and they would otherwise override each others data over and over again.