Removed requirement to invoke event dispatcher when registering/unregistering events
You can now register for listening to events directly
by invoking the static Register/Unregister on the event class itself.

Now that's fancy.

Also: Removed the return value from the Event.Call() method.
it's not exactly useful as such as you'd have to cast on the using end
to get the actual event object you'd expect.