Changed EventDispatcher some more
Changed the thing so that delegates don't need to be cast
everytime when calling them. Casting of event parameters seems
to be unavoidable though, shame.

Performance gains are negligible (round about 2ms on 200k event calls).
But it's more readable now. At least that.
Unsurprisingly, the bottleneck is the TryGetValue() call when
finding the delegates to invoke. Don't think there's much to do about it.

Also did some random stuff in random files and
removed the old SignalProcessorTracker. Because it's not required, really.