Added custom data field & sequence exeuctor can be paused
custom data field ain't really ready yet though.
But it's a start.

Sequence executor can be paused manually or as CSP input func.
It mwill cause the ticking of the behaviour tree to be skipped,
for as long as the pause-toggle flag is set.

You can also unpause the game, of course.